[IMP] account_check_printing_report_base: Remove unnecessary dependency
report_wkhtmltopdf_param is not beeing used so remove it as dependency.

In version 10, this dependency was added because the --disable-smart-shrinking option was being used. However, starting from version 12, this option is no longer necessary, so keeping the dependency is neither required nor correct.

The dependency has been added to the modules that specifically use the flag, in this case, account_check_printing_report_dlt103 and account_check_printing_report_sslm102.